rename interaction files for view count
