+++ components/shared/Votes.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { downvoteAnswer, upvoteAnswer } from "@/lib/actions/answer.action";
-import { viewQuestion } from "@/lib/actions/interaction";
+import { viewQuestion } from "@/lib/actions/interaction.action";
 
 import {
   downvoteQuestion,
