+++ database/interaction.model.ts
